refactor: Move Chip components to chips directory
