+++ dev-client/src/components/SiteCard.tsx
@@ -22,13 +22,13 @@ import {Site} from 'terraso-client-shared/site/siteTypes';
 
 import {IconButton} from 'terraso-mobile-client/components/buttons/icons/IconButton';
 import {Card} from 'terraso-mobile-client/components/Card';
+import {PeopleChip} from 'terraso-mobile-client/components/chips/PeopleChip';
 import {
   Box,
   Heading,
   Row,
   Text,
 } from 'terraso-mobile-client/components/NativeBaseAdapters';
-import {PeopleChip} from 'terraso-mobile-client/components/PeopleChip';
 import {StaticMapView} from 'terraso-mobile-client/components/StaticMapView';
 import {useNavigation} from 'terraso-mobile-client/navigation/hooks/useNavigation';
 import {useSelector} from 'terraso-mobile-client/store';
