+++ dev-client/src/screens/LocationScreens/LocationDashboardContent.tsx
@@ -24,6 +24,7 @@ import {ScrollView} from 'native-base';
 import {SitePrivacy} from 'terraso-client-shared/site/siteTypes';
 import {Coords} from 'terraso-client-shared/types';
 
+import {PeopleChip} from 'terraso-mobile-client/components/chips/PeopleChip';
 import {HelpContentSpacer} from 'terraso-mobile-client/components/content/HelpContentSpacer';
 import {DataPrivacyInfoButton} from 'terraso-mobile-client/components/content/info/privacy/DataPrivacyInfoButton';
 import {
@@ -32,7 +33,6 @@ import {
   Row,
   Text,
 } from 'terraso-mobile-client/components/NativeBaseAdapters';
-import {PeopleChip} from 'terraso-mobile-client/components/PeopleChip';
 import {RadioBlock} from 'terraso-mobile-client/components/RadioBlock';
 import {StaticMapView} from 'terraso-mobile-client/components/StaticMapView';
 import {renderElevation} from 'terraso-mobile-client/components/util/site';
